chore: improve file organization
